requirements.txt
streamlit
google-genai
python-pptx
Pillow